moved websocket part to frontend nodejs server
